+++ Podcast/SearchITunesViewController.swift
@@ -9,55 +9,124 @@
 import UIKit
 import SnapKit
 
-class SearchITunesViewController: ViewController {
+class SearchITunesViewController: ViewController, UITableViewDelegate, UITableViewDataSource, UISearchControllerDelegate, SearchSeriesTableViewDelegate, UISearchBarDelegate {
     
     var searchController: UISearchController!
-    var searchResultsController: SearchITunesTableViewController!
+    var tableView: EmptyStateTableView!
+    
+    var searchResults: [Series] = []
+    let seriesCellIdentifier = "SeriesCell"
+    let seriesCellHeight: CGFloat = 95
     
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .paleGrey
         title = "Search iTunes"
         
-        searchResultsController = SearchITunesTableViewController()
-        view.addSubview(searchResultsController.view)
+        tableView = EmptyStateTableView(frame: view.frame, type: .search)
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.stopLoadingAnimation()
+        tableView.backgroundView?.isHidden = true
+        view.addSubview(tableView)
+        mainScrollView = tableView
         
-        searchController = UISearchController(searchResultsController: searchResultsController)
-
         let topBarAttributes: NSDictionary = [NSAttributedStringKey.foregroundColor: UIColor.sea]
         UIBarButtonItem.appearance().setTitleTextAttributes(topBarAttributes as? [NSAttributedStringKey: Any], for: .normal)
         navigationController?.navigationBar.tintColor = .sea
         
-        searchController.searchBar.showsCancelButton = false
-        searchController.searchBar.showsSearchResultsButton = true
-        searchController.searchBar.searchBarStyle = .minimal
-        searchController.searchBar.placeholder = "Search"
-        searchController.searchBar.delegate = searchResultsController
-        searchController.delegate = searchResultsController
+        searchController = UISearchController(searchResultsController: nil)
+        tableView.tableHeaderView = searchController.searchBar
+        
+        searchController.delegate = self
         searchController.hidesNavigationBarDuringPresentation = false
         searchController.isActive = true
         searchController.obscuresBackgroundDuringPresentation = false
         searchController.dimsBackgroundDuringPresentation = false
+        searchController.searchBar.backgroundColor = .offWhite
+        searchController.searchBar.showsCancelButton = false
+        searchController.searchBar.showsSearchResultsButton = false
+        searchController.searchBar.showsScopeBar = false
+        searchController.searchBar.searchBarStyle = .minimal
+        searchController.searchBar.placeholder = "Search"
+        searchController.searchBar.delegate = self
         definesPresentationContext = true
-
-        if #available(iOS 11.0, *) {
-            navigationItem.searchController = searchController
-        } else {
-            searchResultsController.tableView.tableHeaderView = searchController.searchBar
+    }
+    
+    func fetchData(query: String)  {
+        searchResults = []
+        tableView.reloadData()
+        tableView.startLoadingAnimation()
+        tableView.backgroundView?.isHidden = false
+        
+        System.endpointRequestQueue.cancelAllEndpointRequestsOfType(type: SearchITunesEndpointRequest.self)
+        let request = SearchITunesEndpointRequest(query: query)
+        request.success = { request in
+            guard let results = request.processedResponseValue as? [Series] else { return }
+            if results.isEmpty {
+                self.tableView.backgroundView?.isHidden = false
+            }
+            self.searchResults = results
+            self.tableView.stopLoadingAnimation()
+            self.tableView.reloadData()
         }
+        request.failure = { request in
+            self.tableView.backgroundView?.isHidden = false
+            self.tableView.stopLoadingAnimation()
+        }
+        System.endpointRequestQueue.addOperation(request)
+    }
+    
+    // MARK: TableView
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
+    }
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return seriesCellHeight
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        var cell: SearchSeriesTableViewCell
+        if let dequeuedCell = tableView.dequeueReusableCell(withIdentifier: seriesCellIdentifier) as? SearchSeriesTableViewCell {
+            cell = dequeuedCell
+        } else {
+            cell = SearchSeriesTableViewCell(style: .default, reuseIdentifier: seriesCellIdentifier)
+        }
+        cell.configure(for: searchResults[indexPath.row], index: indexPath.row)
+        cell.delegate = self
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return searchResults.count
+    }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let seriesDetailViewController = SeriesDetailViewController(series: searchResults[indexPath.row])
+        navigationController?.pushViewController(seriesDetailViewController,animated: true)
+    }
+    
+    // MARK: UISearchBarDelegate
+    
+    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
+        guard let searchText = searchBar.text else { return }
+        fetchData(query: searchText)
+    }
+    
+    func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
+        tableView.isHidden = false
+    }
+    
+    // MARK: SearchSeriesCellDelegate
+    
+    func searchSeriesTableViewCellDidPressSubscribeButton(cell: SearchSeriesTableViewCell) {
+        guard let indexPath = tableView.indexPath(for: cell) else { return }
+        let series = searchResults[indexPath.row]
+        series.subscriptionChange(completion: cell.setSubscribeButtonToState)
+    }
 
-    }
-    
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        searchController?.searchBar.isHidden = true
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        searchController?.searchBar.isHidden = false
-    }
-    
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
     }
